src/test/java/com/github/artemkorsakov/digital/PandigitalTests.java
package com.github.artemkorsakov.digital;

import org.testng.Assert;
import org.testng.annotations.Test;

import java.util.Arrays;

public class PandigitalTests {
    @Test
    public void testIsPandigital() {
        Assert.assertTrue(Pandigital.isPandigital(192384576L));
        Assert.assertTrue(Pandigital.isPandigital("192384576"));
        Assert.assertTrue(Pandigital.isPandigital(918273645L));
        Assert.assertTrue(Pandigital.isPandigital("918273645"));
        Assert.assertFalse(Pandigital.isPandigital(192384571L));
        Assert.assertFalse(Pandigital.isPandigital("192384571"));
        Assert.assertFalse(Pandigital.isPandigital(192384570L));
        Assert.assertFalse(Pandigital.isPandigital("192384570"));

        Assert.assertTrue(Pandigital.isPandigital(2134, 4));
        Assert.assertTrue(Pandigital.isPandigital("2134", 4));
        Assert.assertTrue(Pandigital.isPandigital(7652413, 7));
        Assert.assertTrue(Pandigital.isPandigital("7652413", 7));
        Assert.assertFalse(Pandigital.isPandigital(2144, 4));
        Assert.assertFalse(Pandigital.isPandigital("2144", 4));
        Assert.assertFalse(Pandigital.isPandigital(2135, 4));
        Assert.assertFalse(Pandigital.isPandigital("2135", 4));
        Assert.assertFalse(Pandigital.isPandigital(2140, 4));
        Assert.assertFalse(Pandigital.isPandigital("2140", 4));

        Assert.assertTrue(Pandigital.isPandigital(1406357289L, Arrays.asList(9, 8, 7, 6, 5, 4, 3, 2, 1, 0)));
        Assert.assertTrue(Pandigital.isPandigital(140657289L, Arrays.asList(9, 8, 7, 6, 5, 4, 2, 1, 0)));
        Assert.assertTrue(Pandigital.isPandigital("1406357289", Arrays.asList(9, 8, 7, 6, 5, 4, 3, 2, 1, 0)));
        Assert.assertTrue(Pandigital.isPandigital("140657289", Arrays.asList(9, 8, 7, 6, 5, 4, 2, 1, 0)));
        Assert.assertFalse(Pandigital.isPandigital(1406357289L, Arrays.asList(9, 8, 7, 6, 5, 4, 2, 1, 0)));
        Assert.assertFalse(Pandigital.isPandigital(140657289L, Arrays.asList(9, 8, 7, 6, 5, 4, 3, 1, 0)));
        Assert.assertFalse(Pandigital.isPandigital("1406357289", Arrays.asList(9, 8, 7, 6, 5, 4, 2, 1, 0)));
        Assert.assertFalse(Pandigital.isPandigital("140657289", Arrays.asList(9, 8, 7, 6, 5, 4, 3, 1, 0)));
    }
}
